sunbeam-python/requirements.txt
pbr>2.1.0 # Apache 2.0
git+https://github.com/albertodonato/snap-helpers#egg=snap-helpers # LGPLv3
click>=8.1.3 # BSD
rich # MIT

# Used for communication with snapd socket
requests # Apache 2
requests-unixsocket # Apache 2
urllib3<2 # https://github.com/psf/requests/issues/6432

# Used for getting local ip address
netifaces
pyroute2

# Generation of user passwords
pwgen

# Interactive subprocess execution
pexpect

# YAML parsing library
pyyaml>=6.0

# Set upper bound to match Juju 3.2.x series target
# 3.2.3 has introduced a regression
juju>=3.2,<3.2.3

# Used in the launch command to launch an instance
petname

# Lightkube is used to interact with the Kubernetes API
lightkube
lightkube-models

# For plugin repo
jsonschema
GitPython

# maas
python-libmaas
